move useEffect from itemList components to activity and diet screens for reading database when it is initialised
