+++ screens/Activities.js
@@ -1,11 +1,14 @@
 import { Button, StyleSheet, Text, View, SafeAreaView, StatusBar, FlatList} from 'react-native'
-import React, {useLayoutEffect, useState, useContext} from 'react'
+import React, {useLayoutEffect, useState, useContext, useEffect} from 'react'
 import ItemList from '../components/ItemList'
 import { ThemeContext } from '../contexts/ThemeContext';
+import { database } from '../firebase/firebaseSetup';
+import { collection, onSnapshot } from 'firebase/firestore';
 
 // Screen that displays a list of activities
 export default function Activities({ navigation }) {
   const { theme } = useContext(ThemeContext)
+	const [activities, setActivities] = useState([]);
 
   // Add a button to the header that navigates to the Add An Activity screen
 	useLayoutEffect(() => {
@@ -24,10 +27,22 @@ export default function Activities({ navigation }) {
 	}, [navigation])
 
 
+	useEffect(() => {
+		const unsubscribe = onSnapshot(collection(database, 'activities'), (snapshot) => {
+			const activitiesData = snapshot.docs.map(doc => ({
+				id: doc.id,
+				...doc.data()
+			}));
+			setActivities(activitiesData);
+		});
+		return () => unsubscribe();
+  }, []);
+
+
   return (
     <SafeAreaView style={[styles.activitiesContainer, {backgroundColor: theme.backgroundColor}]}>
       {/* Display a list of activities */}
-      <ItemList type='activities'/>
+      <ItemList type='activities' items={activities}/>
     </SafeAreaView>
   )
 }
